renamed LoadingComponents to LoadingState & moved LoadingState to ui/components/shared/
