+++ ui/rtk-query/ability.tsx
@@ -3,7 +3,7 @@ import { useGetUserKeysQuery } from './userKeys';
 import _ from 'lodash';
 import CustomErrorMessage from '@/components/ErrorPage';
 import DefaultError from '@/components/General/error-404';
-import { DynamicFullScreenLoader } from '@/components/LoadingComponents/DynamicFullscreenLoader';
+import { DynamicFullScreenLoader } from '@/components/shared/LoadingState/DynamicFullscreenLoader';
 import {
   useGetProviderCapabilitiesQuery,
   useGetSelectedOrganization,
